refactor to account for file location changes
