+++ frontend/src/components/buttons/index.ts
@@ -1,0 +1,7 @@
+export * from './AddButton/AddButton';
+export * from './AuthenticationButton/AuthenticationButton';
+export * from './Button/Button';
+export * from './CloseButton/CloseButton';
+export * from './EditButton/EditButton';
+export * from './SubmitCancelButton/SubmitCancelButton';
+export * from './ViewButton/ViewButton';
