+++ frontend/src/components/dialogs/index.js
@@ -1,0 +1,6 @@
+export * from './AuthenticationDialog/AuthenticationDialog';
+export * from './Dialog/Dialog';
+export * from './ErrorDialog/ErrorDialog';
+export * from './NewFormDialog/NewForm.Dialog';
+export * from './UpdateFormDialog/UpdateForm.Dialog';
+export * from './ViewFormDialog/ViewForm.Dialog';
