+++ frontend/src/components/forms/index.ts
@@ -1,0 +1,2 @@
+export * from './Form';
+export * from './form.d';
