+++ frontend/src/components/index.ts
@@ -6,12 +6,16 @@ export * from './buttons/Button/Button';
 export * from './dialogs/AuthenticationDialog/AuthenticationDialog';
 export * from './dialogs/Dialog/Dialog';
 export * from './dialogs/ErrorDialog/ErrorDialog';
+export * from './dialogs/NewForm.Dialog';
+export * from './dialogs/UpdateForm.Dialog';
+export * from './dialogs/ViewForm.Dialog';
+export * from './forms/form.d';
 export * from './loading/Loading/Loading';
 export * from './stepper/StatusStepper';
+export * from './table/InquiryTableContainer/InquiryTableContainer';
 export * from './table/TableContainer/TableContainer';
+export * from './table/UserTableContainer/UserTableContainer';
 export * from './tablecells/DataCell/DateCell';
 export * from './tablecells/RolesCell/RolesCell';
 export * from './tablecells/StatusCell/StatusCell';
 export * from './tabs/CustomeTabPanel/CustomTabPanel';
-export * from './table/InquiryTableContainer/InquiryTableContainer';
-export * from './table/UserTableContainer/UserTableContainer';
